requirements.txt
# Core FastAPI dependencies
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18

# AWS Bedrock
boto3==1.35.75
requests==2.32.3

# Data processing
pydantic==2.10.3
pydantic-settings==2.6.1

# Video processing (optional, if analyzing video files)
opencv-python==4.10.0.84

# File handling
aiofiles==24.1.0

# Environment variables
python-dotenv==1.0.1
